ACS-6389: add new common mvn module, refactor and clean-up code
