+++ bulk-ingester/src/integration-test/java/org/alfresco/hxi_connector/bulk_ingester/event/CamelIngestEventPublisherIntegrationTest.java
@@ -26,8 +26,8 @@
 
 package org.alfresco.hxi_connector.bulk_ingester.event;
 
-import static org.alfresco.hxi_connector.bulk_ingester.processor.mapper.AlfrescoNodeMapper.CREATED_AT_PROPERTY;
-import static org.alfresco.hxi_connector.bulk_ingester.processor.mapper.AlfrescoNodeMapper.TYPE_PROPERTY;
+import static org.alfresco.hxi_connector.common.constant.NodeProperties.CREATED_AT_PROPERTY;
+import static org.alfresco.hxi_connector.common.constant.NodeProperties.TYPE_PROPERTY;
 
 import java.io.Serializable;
 import java.util.List;
