+++ live-ingester/src/test/java/org/alfresco/hxi_connector/live_ingester/adapters/messaging/repository/mapper/property/PropertiesMapperTest.java
@@ -33,6 +33,7 @@ import static org.mockito.BDDMockito.given;
 import static org.mockito.Mockito.mock;
 
 import static org.alfresco.hxi_connector.common.constant.NodeProperties.CONTENT_PROPERTY_KEY;
+import static org.alfresco.hxi_connector.common.constant.NodeProperties.NAME_PROPERTY_KEY;
 import static org.alfresco.hxi_connector.live_ingester.domain.usecase.metadata.model.PropertyDelta.deleted;
 import static org.alfresco.hxi_connector.live_ingester.util.TestUtils.mapWith;
 import static org.alfresco.repo.event.v1.model.EventType.NODE_CREATED;
@@ -51,7 +52,6 @@ import org.alfresco.repo.event.v1.model.RepoEvent;
 
 class PropertiesMapperTest
 {
-    private static final String NAME_PROPERTY_KEY = "cm:name";
     PropertiesMapper propertiesMapper = new PropertiesMapper();
 
     @Test
